finished refactoring, cleaned imports of maths package.
