+++ lisa-kernel/src/main/scala/lisa/kernel/fol/EquivalenceChecker.scala
@@ -1,6 +1,5 @@
 package lisa.kernel.fol
 
-import java.lang
 import scala.collection.mutable
 import scala.math.Numeric.IntIsIntegral
