+++ lisa-sets/src/main/scala/lisa/Main.scala
@@ -1,19 +1,19 @@
 package lisa
 
-import lisa.prooflib.BasicMain
 import lisa.SetTheoryLibrary
+import lisa.prooflib.BasicMain
 
 /**
  * The parent trait of all theory files containing mathematical development
  */
 trait Main extends BasicMain {
- 
+
   export lisa.fol.FOL.{*, given}
   export SetTheoryLibrary.{given, _}
   export lisa.prooflib.BasicStepTactic.*
   export lisa.prooflib.SimpleDeducedSteps.*
 
-  export lisa.automation.kernel.Tautology
+  export lisa.automation.Tautology
   export lisa.automation.Substitution
   export lisa.automation.Tableau
 
@@ -23,12 +23,10 @@ trait Main extends BasicMain {
   knownDefs.update(powerSet, Some(powerAxiom))
   knownDefs.update(subset, Some(subsetAxiom))
 
-
   extension (symbol: ConstantLabel[?]) {
     def definition: JUSTIFICATION = {
       getDefinition(symbol).get
     }
   }
-
 
 }
