+++ lisa-sets/src/main/scala/lisa/maths/settheory/orderings/Induction.scala
@@ -1,11 +1,7 @@
-package lisa.maths
-package settheory
-package orderings
+package lisa.maths.settheory.orderings
 
-import lisa.automation.kernel.OLPropositionalSolver.Tautology
-import lisa.automation.kernel.SimplePropositionalSolver.*
 import lisa.automation.settheory.SetTheoryTactics.*
-import lisa.kernel.proof.SequentCalculus as SC
+
 import lisa.maths.Quantifiers.*
 import lisa.maths.settheory.SetTheory.*
 import lisa.maths.settheory.orderings.InclusionOrders.*
@@ -13,38 +9,25 @@ import lisa.maths.settheory.orderings.Ordinals.*
 import lisa.maths.settheory.orderings.PartialOrders.*
 import lisa.maths.settheory.orderings.Segments.*
 import lisa.maths.settheory.orderings.WellOrders.*
-import lisa.prooflib.BasicStepTactic.*
-import lisa.prooflib.Library
-import lisa.prooflib.ProofTacticLib
-import lisa.automation.Substitution
-import lisa.utils.FOLPrinter
 
 object Induction extends lisa.Main {
 
   // var defs
+  private val a = variable
+  private val b = variable
+  private val p = variable
+  private val r = variable
+  private val t = variable
   private val w = variable
   private val x = variable
   private val y = variable
   private val z = variable
-  private val h = formulaVariable
-  private val t = variable
-  private val a = variable
-  private val b = variable
-  private val c = variable
-  private val d = variable
+  
 
   // relation and function symbols
-  private val r = variable
-  private val p = variable
-  private val q = variable
-  private val f = variable
-  private val g = variable
-  private val F = function[1]
-  private val G = function[2]
 
   private val P = predicate[1]
   private val Q = predicate[1]
-  private val schemPred = predicate[1]
 
   /**
    * Theorem --- Well Ordered Induction on a Subclass
