+++ lisa-sets/src/test/scala/lisa/automation/TableauTest.scala
@@ -1,8 +1,9 @@
 package lisa.test.automation
 
-import lisa.automation.Tableau._
+import lisa.SetTheoryLibrary.{_, given}
 import lisa.automation.Substitution.*
-import lisa.settheory.SetTheoryLibrary.{_, given}
+import lisa.automation.Tableau.*
+import lisa.fol.FOL.{_, given}
 import lisa.utils.K.SCProofChecker.checkSCProof
 import lisa.utils.parsing.FOLPrinter.prettyFormula
 import lisa.utils.parsing.FOLPrinter.prettySCProof
@@ -39,16 +40,17 @@ class TableauTest extends AnyFunSuite {
 }
 object TableauTest {
 
-  val w = variable
-  val x = variable
-  val y = variable
-  val z = variable
-
   val a = formulaVariable
   val b = formulaVariable
   val c = formulaVariable
   val d = formulaVariable
   val e = formulaVariable
+
+  val w = variable
+  val x = variable
+  val y = variable
+  val z = variable
+
 
   val f = function[1]
   val g = function[1]
