+++ lisa-sets/src/test/scala/lisa/utilities/SubstitutionTacticTest.scala
@@ -1,10 +1,9 @@
 package lisa.automation
 
-import lisa.test.ProofTacticTestLib
-
+import lisa.automation.Substitution
 import lisa.kernel.proof.RunningTheory
 import lisa.kernel.proof.SequentCalculus as SC
-import lisa.automation.Substitution
+import lisa.test.ProofTacticTestLib
 import lisa.utils.parsing.FOLParser.*
 import lisa.utils.parsing.FOLPrinter.*
 import org.scalatest.funsuite.AnyFunSuite
