+++ lisa-utils/src/main/scala/lisa/utils/Serialization.scala
@@ -650,7 +650,6 @@ object Serialization {
         try {
           Class.forName(obj + "$").getField("MODULE$").get(null)
         } catch { case _ => "Not found: " + obj }
-        Class.forName("lisa.settheory.SetTheoryLibrary" + "$").getField("MODULE$").get(null)
 
         j(0) match
           case 'a' => theory.getAxiom(name).get
